misc items up for v0.1.6
